Add files via upload
The following is a specific tutorial for the EinScan Pro 2X.  This hand held scanner collects dimensional surface data through structured light.  You will also need a PC to connect to the scanner and  collect the real-time data as well as build the 3D model of your scanned object.  Be patient.  If the scanner is not picking up an area of your specimen, go back to an area the scan has already successfully "read."  We found experimentation and patience were key in producing our first successful scans.